feat: add circuit breaker for Qdrant client
